evalscope_perf/requirements.txt
evalscope[perf]
pyyaml